+++ packages/office-ui-fabric-react/src/components/OverflowSet/OverflowSet.doc.tsx
@@ -4,10 +4,12 @@ import { OverflowSetCustomExample } from './examples/OverflowSet.Custom.Example'
 import { IDocPageProps } from '../../common/DocPage.types';
 import { OverflowSetBasicExample } from './examples/OverflowSet.Basic.Example';
 import { OverflowSetVerticalExample } from './examples/OverflowSet.Vertical.Example';
+import { OverflowSetBasicReversedExample } from './examples/OverflowSet.BasicReversed.Example';
 
 const OverflowSetCustomExampleCode = require('!raw-loader!office-ui-fabric-react/src/components/OverflowSet/examples/OverflowSet.Custom.Example.tsx') as string;
 const OverflowSetBasicExampleCode = require('!raw-loader!office-ui-fabric-react/src/components/OverflowSet/examples/OverflowSet.Basic.Example.tsx') as string;
 const OverflowSetVerticalExampleCode = require('!raw-loader!office-ui-fabric-react/src/components/OverflowSet/examples/OverflowSet.Vertical.Example.tsx') as string;
+const OverflowSetBasicReversedExampleCode = require('!raw-loader!office-ui-fabric-react/src/components/OverflowSet/examples/OverflowSet.BasicReversed.Example.tsx') as string;
 
 export const OverflowSetPageProps: IDocPageProps = {
   title: 'OverflowSet',
@@ -29,6 +31,11 @@ export const OverflowSetPageProps: IDocPageProps = {
       title: 'OverflowSet Custom Example',
       code: OverflowSetCustomExampleCode,
       view: <OverflowSetCustomExample />
+    },
+    {
+      title: 'OverflowSet Reversed Example',
+      code: OverflowSetBasicReversedExampleCode,
+      view: <OverflowSetBasicReversedExample />
     }
   ],
   overview: require<string>('!raw-loader!office-ui-fabric-react/src/components/OverflowSet/docs/OverflowSetOverview.md'),